Integrate core sdk to tg bot
