+++ packages/tg-bot/src/services/factory.ts
@@ -1,0 +1,6 @@
+import { Factory } from 'core'
+import { constants } from 'starknet'
+
+import { provider } from './provider'
+
+export const factory = new Factory({ provider, chainId: constants.StarknetChainId.SN_MAIN })
