--- packages/tg-bot/src/utils/ekubo.ts
@@ -1,6 +1,0 @@
-import { EKUBO_TICK_SIZE, EKUBO_TICK_SIZE_LOG, EKUBO_TICK_SPACING } from './constants'
-
-export const getInitialPrice = (startingTick: number) => EKUBO_TICK_SIZE ** startingTick
-
-export const getStartingTick = (initialPrice: number) =>
-  Math.floor(Math.log(initialPrice) / EKUBO_TICK_SIZE_LOG / EKUBO_TICK_SPACING) * EKUBO_TICK_SPACING
